Build-159: New HUD
# Changes
Add In Game Hud.
Add Point at center of screen.
Can hide hud by pressing F1.
Reduced Water wave effect.

# BugFixes

+++ src/main/java/net/guerra24/voxel/client/resources/models/Tessellator.java
@@ -398,8 +398,8 @@ public class Tessellator {
 
 	}
 
-	public void generateCube(int x, int y, int z, int size, boolean top, boolean bottom, boolean left, boolean right,
-			boolean front, boolean back, IBlock block, float light) {
+	public void generateCube(float x, float y, float z, float size, boolean top, boolean bottom, boolean left,
+			boolean right, boolean front, boolean back, IBlock block, float light) {
 		int id = block.getId();
 		float l = light / 15f;
 		if (id != Block.Air.getId()) {
